+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.6
 * Add FetchBuilder.initBuilder param (to be used with fetchAtInit false)
+* New static AsyncTaskBuilder.runTask method that allows to run headless task safely
 
 ## 0.0.5
 * Default FetchBuilderErrorWidget now handles isDense parameter